+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.1.9a
+- New module: 'domain_trusts'
+- New module: 'computer <target>'
+- New module: 'domain_sid'
+- Added SID translator (handlers/sid_translate.py)
+- Now to add a member to a group that have spaces on the name, you need to use % as the space, for example: Domain%20Admins (if we use _ for example, we cannot add a user for a group with _ in the name, like IIS_IUSRS)
+- Improved help command (help, help <protocol_name>)
+- Module 'kerberoasting' now return kerberoastable user TGS
+
 # 1.1.9
 - New module: 'change_password <username> <new_password>'
 - Added tab "Admin Privileges" to help table